requirements.txt
gpiozero
paho-mqtt
flask